+++ .gitignore
@@ -1,3 +1,6 @@
+# Mine
+uploads/
+
 # Logs
 logs
 *.log
