+++ package.json
@@ -19,6 +19,7 @@
   "dependencies": {
     "bcrypt-nodejs": "^0.0.3",
     "cookie-parser": "^1.4.5",
+    "dotenv": "^10.0.0",
     "express": "^4.17.1",
     "express-session": "^1.17.1",
     "express-validator": "^5.3.1",
