+++ public/js/script.js
@@ -24,18 +24,18 @@
         // SEMAFORO
         const proyectos = document.querySelectorAll('.proyecto a');
         proyectos.forEach(proyecto => {
-            const fecha_inicio = Date.parse(proyecto.childNodes[4].innerText); // La fecha de inicio se parsea a milisegundos
-            const fecha_entrega = Date.parse(proyecto.childNodes[5].innerText); // La fecha de entrega se parsea a milisegundos
-            const diferenciaInicioEntrega = (fecha_entrega - fecha_inicio) / 86400000; // Diferencia en dias entre las fechas
-            const fechaActual = Date.now(); // Fecha actual en milisegundos
+            const fecha_inicio = Date.parse(proyecto.childNodes[4].innerText);
+            const fecha_entrega = Date.parse(proyecto.childNodes[5].innerText);
+            const diferenciaInicioEntrega = (fecha_entrega - fecha_inicio) / 86400000;
+            const fechaActual = Date.now();
             const diferenciaActualEntrega = (fecha_entrega - fechaActual) / 86400000;
-            if (fechaActual >= fecha_entrega) { // Fechas en milisegundos para hacer la resta directamente
+            if (fechaActual >= fecha_entrega) {
                 /* RED STATUS */
                 proyecto.childNodes[6].childNodes[3].style.opacity = "1";
             } else if (diferenciaActualEntrega <= (diferenciaInicioEntrega / 2)) {
                 /* YELLOW STATUS */
                 proyecto.childNodes[6].childNodes[2].style.opacity = "1";
-            } else { // Ni una ni otra
+            } else {
                 /* GREEN STATUS */
                 proyecto.childNodes[6].childNodes[1].style.opacity = "1";
             }
@@ -49,49 +49,55 @@
             1 dia = 24 horas = 86400000 milisegundos
         */
         // CONTEO
-        const totalTareas = document.querySelectorAll('.tarea').length;
+        const totalTareas = document.querySelectorAll('.tarea');
         const areas = document.querySelector('input#areas');
-        if (totalTareas > 0) {
+        if (totalTareas.length > 0) {
             conteoyPorcentaje(totalTareas, areas);
         }
 
         function conteoyPorcentaje(totalTareas, areas) {
             resetearBarras();
-            if (totalTareas == 0) {
+            if (totalTareas.length == 0) {
                 document.querySelector('h3#null-tareas').style.display = "block";
             }
-            const departamentosString = areas.value.split(',');
             const tareasArray = []; // Arreglo para guardar distintos departamentos con sus tareas relacionadas
-            departamentosString.forEach(departamento => {
-                tareasArray.push(document.querySelectorAll(`.tarea[data-departamento=${departamento}]`));
-            });
-            var totalCompletadas = 0;
-            var totalNoCompletadas = 0;
+            if (areas.value != '') {
+                const departamentosString = areas.value.split(',');
+                departamentosString.forEach(departamento => {
+                    tareasArray.push(document.querySelectorAll(`.tarea[data-departamento=${departamento}]`));
+                });
+            }
+            let totalCompletadas = 0;
+            let totalNoCompletadas = 0;
             tareasArray.forEach(tareasColeccion => {
-                var totalTareasDepartamento = 0;
-                var tareasCompletadasDepartamento = 0;
+                let totalTareasDepartamento = 0;
+                let tareasCompletadasDepartamento = 0;
                 tareasColeccion.forEach(tarea => {
                     totalTareasDepartamento += 1;
                     if (tarea.children[4].children[0].classList.contains('activo')) {
-                        totalCompletadas += 1;
-                        tareasCompletadasDepartamento += 1
-                    } else {
-                        totalNoCompletadas += 1;
+                        tareasCompletadasDepartamento += 1;
                     }
-
                     document.querySelector(`.barraDepto[data-departamento=${tarea.dataset.departamento}]`).childNodes[1].childNodes[0].style.width = `${(tareasCompletadasDepartamento / totalTareasDepartamento).toFixed(2)*100}%`;
                     document.querySelector(`.barraDepto[data-departamento=${tarea.dataset.departamento}]`).childNodes[2].innerText = `${tareasCompletadasDepartamento}/${totalTareasDepartamento}`;
                 })
+            });
+            totalTareas.forEach(tarea => {
+                if (tarea.children[4].children[0].classList.contains('activo')) {
+                    totalCompletadas += 1;
+                } else {
+                    totalNoCompletadas += 0;
+                }
             })
 
-            document.querySelector('.detallesproyecto').childNodes[1].innerText = `Cantidad de tareas: ${totalTareas}`;
+            document.querySelector('.detallesproyecto').childNodes[1].innerText = `Cantidad de tareas: ${totalTareas.length}`;
             document.querySelector('.detallesproyecto').childNodes[2].innerText = `Tareas terminadas: ${totalCompletadas}`;
             document.querySelector('.detallesproyecto').childNodes[3].innerText = `Tareas pendientes: ${totalNoCompletadas}`;
 
-            if (totalTareas == 0) {
-                var porcentajeTotal = 0;
+            let porcentajeTotal;
+            if (totalTareas.length == 0) {
+                porcentajeTotal = 0;
             } else {
-                var porcentajeTotal = ((totalCompletadas / totalTareas).toFixed(2)) * 100;
+                porcentajeTotal = ((totalCompletadas / totalTareas.length).toFixed(2)) * 100;
             }
             const circuloProgreso = document.querySelector('.circulo');
             circuloProgreso.firstChild.innerHTML = porcentajeTotal.toFixed(0) + "%";
@@ -111,9 +117,8 @@
             })
         }
 
-        // Bloque para realizar las peticiones a Axios y DOMScripting
+        // Axios y DOMScripting
         if (totalTareas) {
-            // Seleccionar los iconos para completar o descompletar las tareas
             const iconosTarea = document.querySelectorAll('.iconos-tarea');
             iconosTarea.forEach(tarea => {
                 tarea.addEventListener('click', e => {
@@ -121,7 +126,6 @@
                     if (e.target.classList.contains('fa-check-circle')) { // Al completar
                         const icono = e.target;
                         const idTarea = icono.parentElement.parentElement.parentElement.dataset.tarea;
-                        // Request hacia /tareas/:id
                         const url = `${location.origin}/tarea-completada/${idTarea}`;
                         // No pasamos parametros como tal, sino indicamos la url donde se hara el patch
                         axios.patch(url, { idTarea })
@@ -129,36 +133,32 @@
                                 if (response.status == 200) {
                                     icono.parentElement.classList.add("activo");
                                     icono.parentElement.nextElementSibling.classList.remove("activo");
-                                    const totalTareas = document.querySelectorAll('.tarea').length;
+                                    const totalTareas = document.querySelectorAll('.tarea');
                                     conteoyPorcentaje(totalTareas, areas);
                                 }
                             })
                     } else if (e.target.classList.contains('fa-times-circle')) { // Al descompletar
                         const icono = e.target;
                         const idTarea = icono.parentElement.parentElement.parentElement.dataset.tarea;
-                        // Request hacia /tareas/:id
                         const url = `${location.origin}/tarea-descompletada/${idTarea}`;
-                        // Indicar la url del patch y el ID
                         axios.patch(url, { idTarea })
                             .then(function(response) {
                                 if (response.status == 200) {
                                     icono.parentElement.classList.add("activo");
                                     icono.parentElement.previousElementSibling.classList.remove("activo");
-                                    const totalTareas = document.querySelectorAll('.tarea').length;
+                                    const totalTareas = document.querySelectorAll('.tarea');
                                     conteoyPorcentaje(totalTareas, areas);
                                 }
                             })
                     }
                 });
             });
-            // Seleccionar los tres puntos de cada tarea para mostrar las opciones de editar o eliminar
             const opcionesTarea = document.querySelectorAll('i.fa-ellipsis-v');
             opcionesTarea.forEach(opcion => {
                 opcion.addEventListener('click', e => {
                     if (e.target.classList.contains('eliminar-tarea')) { // Al eliminar
                         const accion = e.target;
                         const idTarea = accion.parentElement.parentElement.parentElement.dataset.tarea;
-                        // Request hacia /tareas/:id
                         const url = `${location.origin}/tarea-eliminar/${idTarea}`;
                         swal({
                                 title: "¿Eliminar tarea?",
@@ -177,7 +177,7 @@
                                         .then(function(response) {
                                             if (response.status == 200) {
                                                 accion.parentElement.parentElement.parentElement.remove();
-                                                const totalTareas = document.querySelectorAll('.tarea').length;
+                                                const totalTareas = document.querySelectorAll('.tarea');
                                                 conteoyPorcentaje(totalTareas, areas);
                                             }
                                         });
@@ -239,7 +239,6 @@
                     })
                     .then((result) => {
                         if (result.value == true) {
-                            // No pasamos parametros como tal, sino indicamos la url donde se hara el patch
                             axios.delete(url, { urlProyecto })
                                 .then(function(response) {
                                     if (response.status == 200) {
