+++ variables.env
@@ -1,0 +1,7 @@
+DATABASE=workflow
+HOST=127.0.0.1
+PORT=3307
+USER=root
+PASSWORD=root
+SECRET=itsasecret
+KEY=secretkey
