+++ views/formularioRegistro.pug
