+++ views/login.pug
@@ -17,7 +17,6 @@ html(lang='es')
             i.fas.fa-lock(style="color:#fbfbfb")
           if mensaje
             p(style="color:var(--correcto)") #{mensaje}
-          //- i.fas.fa-unlock
         form(action='/iniciar-sesion' method="POST")
           .campo
             label(for='email') E-mail:
@@ -32,6 +31,6 @@ html(lang='es')
         if failedLogin
           a(href="/reestablecer-password" style="color:#8de4af") Olvidé mi contraseña D:
       .contenido
-        img(src=`../${fondo}` alt="El gif")
+        img(src=`/${fondo}` alt="El gif")
     
     include partials/footer
