+++ views/proyecto.pug
@@ -65,9 +65,6 @@ body
                     a.nocompleto.tooltip(href='#')
                       i.far.fa-times-circle
                        span.tooltiptext Descompletar
-                  //- a.basura.tooltip(href='#')
-                  //-   i.far.fa-trash-alt
-                  //-     span.tooltiptext Eliminar
                 i.fas.fa-ellipsis-v.tooltip
                   .tooltiptext
                     a(href=`/proyecto/${proyecto.url}/editar-tarea/${tarea.id_tarea}`) Editar
